requirements.txt
streamlit
faiss-cpu
gitpython
sentence-transformers
langchain
langchain-community
ollama
unstructured
